Initial mechanic done. Need to review code - not confident it is written optimially yet

+++ tutorials/Unity/ui-toolkit-grid-inventory/finalProject/Assets/Scripts/ItemDefinition.cs
@@ -17,6 +17,6 @@ public class ItemDefinition : ScriptableObject
 [Serializable]
 public struct Dimensions
 {
-    public int Height;
-    public int Width;
+    public float Height;
+    public float Width;
 }
